# performance: fix tzOffset cache not working for UTC offset strings