### Problem
When passing a UTC offset string like `+00:00` to the `tzOffset` function, the cache is not used properly.
This causes significant performance degradation on Android devices 😓 

### Benchmark

#### `new TZDate()`
```
const start = new Date().getTime();
new TZDate(new Date(), "+00:00");
const end = new Date().getTime();

console.log(end - start);

// iOS (React Native):      iPhone 11 -> 0ms
// Android (React Native):  Pixel 7a -> 20ms, Galaxy A21 SC-42A -> 7ms, Redmi 12 -> 65ms
//                          (After modification, all 0ms.)
```

#### `Intl.DateTimeFormat` 
```
const start = new Date().getTime();
[...Array(1000)].forEach(() => {
  try {
    new Intl.DateTimeFormat("en-GB", {
      timeZone: "+00:00",
      hour: "numeric",
      timeZoneName: "longOffset",
    }).format;
  } catch {}
});
const end = new Date().getTime();

console.log(end - start);

// Chrome:                  MacBook Pro 2020 -> 152ms
// iOS (React Native):      iPhone 11 -> 20ms
// Android (React Native):  Pixel 7a -> 5087ms, Galaxy A21 SC-42A -> 2027ms, Redmi 12 -> 5098ms
```